backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
pytest==8.0.2
httpx==0.27.0
python-dotenv==1.0.1
